Extract NFT card image to dedicated file
